employe changed and attendance report added

+++ database/migrations/2025_03_01_101932_add_foreign_key_to_salary_cards.php
@@ -23,11 +23,23 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('salary_cards', function (Blueprint $table) {
-            $table->dropForeign(['employee_id', 'created_by', 'updated_by']);
-            $table->dropColumn('employee_id');
-            $table->dropColumn('created_by');
-            $table->dropColumn('updated_by');
-        });
+        if (Schema::hasTable('salary_cards')) {
+            Schema::table('salary_cards', function (Blueprint $table) {
+                if (Schema::hasColumn('salary_cards', 'employee_id')) {
+                    $table->dropForeign(['employee_id']);
+                    $table->dropColumn('employee_id');
+                }
+
+                if (Schema::hasColumn('salary_cards', 'created_by')) {
+                    $table->dropForeign(['created_by']);
+                    $table->dropColumn('created_by');
+                }
+
+                if (Schema::hasColumn('salary_cards', 'updated_by')) {
+                    $table->dropForeign(['updated_by']);
+                    $table->dropColumn('updated_by');
+                }
+            });
+        }
     }
 };
